Refactored client management
