+++ client/cancel_download_file.go
@@ -1,0 +1,5 @@
+package client
+
+func (context *Context) CancelDownloadFile(fileId int32) {
+	_, _ = context.Client.CancelDownloadFile(fileId, false)
+}
